+++ khal/controllers.py
@@ -187,9 +187,9 @@ def get_events_between(
         # to understand what's going on here this way
         if notstarted:
             if event.allday and event.start < original_start.date():
-                    continue
+                continue
             elif not event.allday and event.start_local < original_start:
-                    continue
+                continue
         if seen is not None and event.uid in seen:
             continue
 
@@ -355,6 +355,8 @@ def new_from_args(collection, calendar_name, conf, dtstart=None, dtend=None,
                   categories=None, repeat=None, until=None, alarms=None,
                   timezone=None, format=None, env=None):
     """Create a new event from arguments and add to vdirs"""
+    if isinstance(categories, str):
+        categories = list([category.strip() for category in categories.split(',')])
     try:
         event = utils.new_event(
             locale=conf['locale'], location=location, categories=categories,
@@ -495,7 +497,10 @@ def edit_event(event, collection, locale, allow_quit=False, width=80):
             value = prompt(question, default)
             if allow_none and value == "None":
                 value = ""
-            getattr(event, "update_" + attr)(value)
+            if attr == 'categories':
+                getattr(event, "update_" + attr)(list([cat.strip() for cat in value.split(',')]))
+            else:
+                getattr(event, "update_" + attr)(value)
             edited = True
 
         if edited:
@@ -515,9 +520,9 @@ def edit(collection, search_string, locale, format=None, allow_past=False, conf=
     for event in events:
         if not allow_past:
             if event.allday and event.end < now.date():
-                    continue
+                continue
             elif not event.allday and event.end_local < now:
-                    continue
+                continue
         event_text = textwrap.wrap(event.format(format, relative_to=now), term_width)
         echo(''.join(event_text))
         if not edit_event(event, collection, locale, allow_quit=True, width=term_width):