+++ CHANGELOG.rst
@@ -7,6 +7,13 @@ Package maintainers and users who have to manually update their installation
 may want to subscribe to `GitHub's tag feed
 <https://github.com/geier/khal/tags.atom>`_.
 
+
+0.10.1
+======
+2019-03-30
+
+* FIX error with the new color priority system and `discover` calendar type
+
 0.10.0
 ======
 2019-03-25